+++ tests/testthat/test-enum.R
@@ -6,9 +6,9 @@ library(dplyr, warn.conflicts=F)  # for data_frame
 source("utils.R")
 
 test_that("reading & writing enum columns", {
-  writeReadTest(as.data.frame(data_frame(x=as.factor(c("foo","bar","baz","foo","baz")))))
+  writeReadTest(as.data.frame(tibble(x=as.factor(c("foo","bar","baz","foo","baz")))))
 })
 
 test_that("nullable enum column", {
-  writeReadTest(as.data.frame(data_frame(x=as.factor(c("foo",NA,"baz","foo","baz",NA)))))
+  writeReadTest(as.data.frame(tibble(x=as.factor(c("foo",NA,"baz","foo","baz",NA)))))
 })