+++ tests/testthat/test-int64.R
@@ -6,6 +6,6 @@ library(dplyr, warn.conflicts=F)  # for data_frame
 source("utils.R")
 
 test_that("reading & writing int64", {
-  writeReadTest(as.data.frame(data_frame(x=bit64::as.integer64(c("9007199254740993")))))
+  writeReadTest(as.data.frame(tibble(x=bit64::as.integer64(c("9007199254740993")))))
 })